+++ src/model.ml
@@ -377,6 +377,8 @@ type ('id, 'term) mterm_node  =
   | Mlength           of 'term
   | Misnone           of 'term
   | Missome           of 'term
+  | Misnat            of 'term
+  | Mtonat            of 'term
   | Moptget           of 'term
   | Mrequiresome      of 'term * 'term
   | Mfloor            of 'term
@@ -1588,6 +1590,8 @@ let cmp_mterm_node
     | Mlength x1, Mlength x2                                                           -> cmp x1 x2
     | Misnone x1, Misnone x2                                                           -> cmp x1 x2
     | Missome x1, Missome x2                                                           -> cmp x1 x2
+    | Misnat x1, Misnat x2                                                             -> cmp x1 x2
+    | Mtonat x1, Mtonat x2                                                             -> cmp x1 x2
     | Moptget x1, Moptget x2                                                           -> cmp x1 x2
     | Mrequiresome (x1, y1), Mrequiresome (x2, y2)                                     -> cmp x1 x2 && cmp y1 y2
     | Mfloor x1, Mfloor x2                                                             -> cmp x1 x2
@@ -2031,6 +2035,8 @@ let map_term_node_internal (fi : ident -> ident) (g : 'id -> 'id) (ft : type_ ->
   | Mlength x                      -> Mlength (f x)
   | Misnone x                      -> Misnone (f x)
   | Missome x                      -> Missome (f x)
+  | Misnat x                       -> Misnat (f x)
+  | Mtonat x                       -> Mtonat (f x)
   | Moptget x                      -> Moptget (f x)
   | Mrequiresome (x, y)            -> Mrequiresome (f x, f y)
   | Mfloor x                       -> Mfloor (f x)
@@ -2471,6 +2477,8 @@ let fold_term (f : 'a -> ('id mterm_gen) -> 'a) (accu : 'a) (term : 'id mterm_ge
   | Mlength x                             -> f accu x
   | Misnone x                             -> f accu x
   | Missome x                             -> f accu x
+  | Misnat x                              -> f accu x
+  | Mtonat x                              -> f accu x
   | Moptget x                             -> f accu x
   | Mrequiresome (x, y)                   -> f (f accu x) y
   | Mfloor x                              -> f accu x
@@ -3485,6 +3493,14 @@ let fold_map_term
   | Missome x ->
     let xe, xa = f accu x in
     g (Missome xe), xa
+
+  | Misnat x ->
+    let xe, xa = f accu x in
+    g (Misnat xe), xa
+
+  | Mtonat x ->
+    let xe, xa = f accu x in
+    g (Mtonat xe), xa
 
   | Misnone x ->
     let xe, xa = f accu x in