+++ src/typing.ml
@@ -1337,8 +1337,12 @@ let coreops : opinfo list =
   @ (List.map
        (fun x -> op "to_string" A.Ctostring `Total None [x] (`Ty A.vtstring) Mint.empty)
        [A.vtnat])
-  @ [op "date_from_timestamp" A.CdateFromTimestamp `Total None [A.vtint] (`Ty A.vtdate) Mint.empty;
-     op "mutez_to_nat" A.CmutezToNat `Total None [A.vtcurrency] (`Ty A.vtnat) Mint.empty]
+  @ [
+    op "isnat"  A.Cisnat `Total None [A.vtint] (`Ty (A.Toption A.vtnat)) Mint.empty;
+    op "to_nat" A.Ctonat `Total None [A.vtint] (`Ty A.vtnat) Mint.empty;
+    op "date_from_timestamp" A.CdateFromTimestamp `Total None [A.vtint] (`Ty A.vtdate) Mint.empty;
+    op "mutez_to_nat" A.CmutezToNat `Total None [A.vtcurrency] (`Ty A.vtnat) Mint.empty
+  ]
 
 
 (* -------------------------------------------------------------------- *)